refactor(llm): extract MemGuard (CMM/DDR guard engine) from LLM::Impl
Move the pre-load memory guard's measurement/extrapolation engine + its state
out of the 5.5k-line LLM.cpp God object into src/runner/MemGuard.hpp:
  - CMM/DDR remaining queries (device_remaining_cmm_mb, cmm_total_remaining_mb)
  - CMM teardown-balance sentry (capture_sentry_baseline / CheckCmmBalance)
  - running load guard: state (rl_* maps, guard_settled_), GuardVerdict, and the
    pure running_guard_eval() extrapolation
MemGuard holds a const LLMAttrType&. Behavior-preserving: the topology-aware
orchestration stays in Impl -- mem_preflight unchanged; running_guard_init()
builds the per-layer device list (layer_devid_for) and delegates; running_guard_
check() delegates to eval + is/mark_guard_settled; guard_device_load/host_load
keep set_last_error in Impl. No load-loop call sites changed.

Verified:
  - compiles with real build_x86 (AXCL) + build_650 (aarch64) flags (EXIT 0)
  - BEHAVIOR verified on AX650N (.47) via tools/test_teardown_cmm_leak.sh with
    qwen3-0.6b: forced mid-load abort (AXLLM_TEST_ABORT_AFTER_LAYER) still fires
    through running_guard_eval; cmm-sentry (CheckCmmBalance) reports CMM balanced,
    delta 0 MB, no leak. PASS.
